# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Elegant Sip API",
    "version": "0.1.0",
    "description": "Backend for the Elegant Sip storefront — single-origin Darjeeling tea.\n\n**Money is integer paise** (1 rupee = 100 paise). Razorpay settles in paise and\nfloats cannot represent money. Totals are always a whole number of rupees:\nGST is rounded with the same rule the storefront uses, so the figure a customer\nis shown and the figure they are charged cannot drift apart.\n\nErrors share one shape across every endpoint, modelled on RFC 9457. `detail` is\nalways safe to display; `requestId` ties a response to the server logs.",
    "contact": {
      "name": "Elegant Sip",
      "email": "elegantsipdarjeeling@gmail.com"
    }
  },
...
  "paths": {
...
    "/v1/products/{slug}": {
      "get": {
        "summary": "Get one product",
        "tags": [
          "Catalogue"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 80,
              "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
            },
            "in": "path",
            "name": "slug",
            "required": true,
            "description": "URL-safe identifier, e.g. \"first-flush-whole-leaf\""
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    "slug": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 80,
                      "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$",
                      "description": "URL-safe identifier, e.g. \"first-flush-whole-leaf\""
                    },
                    "name": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "active",
                        "coming-soon"
                      ],
                      "description": "Coming-soon products are listed but cannot be purchased and carry no price"
                    },
                    "fromPrice": {
                      "anyOf": [
                        {
                          "type": "integer",
                          "minimum": 0,
                          "maximum": 9007199254740991,
                          "description": "Amount in paise (1 rupee = 100 paise)"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "description": {
                      "type": "string"
                    },
                    "longDescription": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "imageSrc": {
                      "type": "string"
                    },
                    "category": {
                      "type": "string"
                    },
                    "tastingNotes": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "bodyLevel": {
                      "anyOf": [
                        {
                          "type": "integer",
                          "minimum": 1,
                          "maximum": 5
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "harvestLabel": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "variants": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid",
                            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                          },
                          "size": {
                            "type": "string",
                            "description": "Display label, e.g. \"Classic · 100 g\""
                          },
                          "price": {
                            "type": "integer",
                            "minimum": 0,
                            "maximum": 9007199254740991,
                            "description": "Amount in paise (1 rupee = 100 paise)"
                          },
                          "stock": {
                            "type": "integer",
                            "minimum": 0,
                            "maximum": 9007199254740991,
                            "description": "Units available in the current lot"
                          },
                          "sku": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "size",
                          "price",
                          "stock",
                          "sku"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "origin": {
                      "anyOf": [
                        {
                          "type": "object",
                          "properties": {
                            "origin": {
                              "type": "string",
                              "description": "e.g. \"Darjeeling, West Bengal, India\""
                            },
                            "estate": {
                              "type": "string",
                              "description": "Must match a garden name — the two are cross-linked"
                            },
                            "elevation": {
                              "type": "string"
                            },
                            "harvest": {
                              "type": "string"
                            },
                            "cultivar": {
                              "type": "string"
                            }
                          },
                          "required": [
                            "origin",
                            "estate",
                            "elevation",
                            "harvest",
                            "cultivar"
                          ],
                          "additionalProperties": false
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "flavorProfile": {
                      "anyOf": [
                        {
                          "type": "object",
                          "properties": {
                            "strength": {
                              "type": "integer",
                              "minimum": 1,
                              "maximum": 5
                            },
                            "astringency": {
                              "type": "integer",
                              "minimum": 1,
                              "maximum": 5
                            },
                            "sweetness": {
                              "type": "integer",
                              "minimum": 1,
                              "maximum": 5
                            },
                            "floral": {
                              "type": "integer",
                              "minimum": 1,
                              "maximum": 5
                            },
                            "caffeine": {
                              "type": "integer",
                              "minimum": 1,
                              "maximum": 5
                            }
                          },
                          "required": [
                            "strength",
                            "astringency",
                            "sweetness",
                            "floral",
                            "caffeine"
                          ],
                          "additionalProperties": false
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "brewingGuide": {
                      "anyOf": [
                        {
                          "type": "object",
                          "properties": {
                            "temperature": {
                              "type": "string"
                            },
                            "time": {
                              "type": "string"
                            },
                            "steeps": {
                              "type": "string"
                            },
                            "leafAmount": {
                              "type": "string"
                            },
                            "notes": {
                              "type": "string"
                            }
                          },
                          "required": [
                            "temperature",
                            "time",
                            "steeps",
                            "leafAmount",
                            "notes"
                          ],
                          "additionalProperties": false
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "rating": {
                      "type": "object",
                      "properties": {
                        "average": {
                          "type": "number",
                          "minimum": 0,
                          "maximum": 5
                        },
                        "count": {
                          "type": "integer",
                          "minimum": 0,
                          "maximum": 9007199254740991
                        }
                      },
                      "required": [
                        "average",
                        "count"
                      ],
                      "additionalProperties": false,
                      "description": "Derived from published reviews. count 0 means no stars are shown at all"
                    }
                  },
                  "required": [
                    "id",
                    "slug",
                    "name",
                    "status",
                    "fromPrice",
                    "description",
                    "longDescription",
                    "imageSrc",
                    "category",
                    "tastingNotes",
                    "bodyLevel",
                    "harvestLabel",
                    "variants",
                    "origin",
                    "flavorProfile",
                    "brewingGuide",
                    "rating"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "description": "Stable machine-readable slug, e.g. \"validation_error\", \"out_of_stock\""
                    },
                    "title": {
                      "type": "string",
                      "description": "Short human-readable summary"
                    },
                    "status": {
                      "type": "integer",
                      "minimum": 400,
                      "maximum": 599
                    },
                    "detail": {
                      "type": "string",
                      "description": "What went wrong and what to do about it — displayable as-is"
                    },
                    "errors": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      }
                    },
                    "requestId": {
                      "type": "string",
                      "description": "Correlates this response with the server logs"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "requestId"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local development"
    },
    {
      "url": "https://api.elegantsip.com",
      "description": "Production"
    }
  ],
...
}